@heanlan 2021-07-23
> Started the review before you made the PR into draft. Take your time with the comments.

Sorry about suddenly making it into draft. I was having some issues with setting-up e2e tests locally last night, so pushed and hope to use the github CI testings during the night. Didn't realize it won't trigger tests without doing rebase. I'll try to fix the local problems now and also do the rebase before making a new commit.