@heanlan 2021-08-03
> Please run the corresponding performance unit tests in connections and exporter pkg to see if there are any changes.

will do