@heanlan 2021-08-10
> For example, why is the priority queue part of the connection store? 
Because we need to modify pq in AddOrUpdateConn methods in both conntrack_connections.go and deny_connections.go. If we only put pq as part of flowExporter struct, these two connection callers doesn't have access to flowExporter.expirePriorityQueue.

> The fact that you have to make `ExpirePriorityQueue` a _public_ field in the connection store, that other modules are accessing directly, is a red flag to me. 
This is my mistake, it can be changed to a private field. Forgot to changed it back.

> Similarly, the fact that the PQ item needs to store all these fields is also a red flag. The PQ does not need to be aware of byte & packet counters.
Srikar suggested in [comment](https://github.com/antrea-io/antrea/pull/2360#discussion_r673360103) that we'd better to move all stats metadata to pqItem, so that exporter go routine will not access connection map. Overall to avoid deadlock, causing by multiple go routine (polling go routine in conntrack connection, export go routine in exporter) have access to multiple resources (priority queue, connection map).
I feel if we make both go routine access connection map lock first, and access priority queue lock later, accessing the resources following the same order can avoid deadlock?

> It may be a good idea to draw some sort of class diagram somewhere to have a clear picture of how the different modules should interact.
Will do it right now

> Also curious about the performance improvements; I see that Srikar also asked about that.
Work got delayed by PTO. Will do.