@heanlan 2021-08-10
having a diagram here:
https://onevmw-my.sharepoint.com/:b:/g/personal/hanlan_vmware_com/ESimBNPRNCdJpBieawY7LFIBBNABRqlbZPZl6DNeqkpYCw?e=ouUvt7