@heanlan 2021-08-11
Running exporter perf test - BenchmarkExportConntrackConns:

working branch:
|conns | ns/op | B/op | allocs/op|
| -------|------ | -----|------- |
|20000 | 1718657 | 7666 | 123|
|30000 | 2497823 | 8505 | 144|
|40000 | 1180317 | 14332 | 237|

main branch:
|conns | ns/op | B/op | allocs/op|
| -------|------ | -----|------- |
|20000 | 10320774 | 1651370 | 9170 |
|30000 | 17695434 | 1358809 | 22884 |
|40000 | 47754824 | 9045501 | 75455 |

The results of BenchmarkExportDenyConns are very close to conntrack conn.

Will try to generate a diagram based on @antoninbas 's proposed design.