@heanlan 2021-08-12
A diagram based on @antoninbas 's proposal: [link](https://drive.google.com/file/d/1VRc_-NTvI3diVunHE5lbAPWk2ZO5fORH/view?usp=sharing). The details logic in grey can mostly be skipped for viewers.

In general, I feel the the separation between connection store and exporter is much clearer. As the fields of each data struct are more relevant. Considering the usage of locks now, connection store only acquire CS lock; while both routines of exporter acquire for CS and PQ's lock. I listed two questions on the diagram, would like to hear opinions @srikartati @antoninbas @zyiou thanks