@antoninbas 2021-08-12
@heanlan for your questions:

1. yes, please pass all the necessary information through the channel to avoid locking. Ideally the locking events would match what I described in my review comment as it has minimal contention IMO.

2. yes also. When I wrote "maps it to a PQ item" I did mean that you needed a new `map` object for this.