@srikartati 2021-08-16
Thanks @heanlan for the detailed design links for the existing approach and the current approach.

> Maybe it was a mistake to separate the connection store and the exporter like this. If all the state for a given connection was stored in one object, the code would be simpler? Curious to hear @srikartati's thoughts on this. If I recall correctly, the original thinking was that there could be too much contention if the 2 functionalities were merged, but a lot have changed since in how we export records.

Yes maintaining one object definitely crossed my mind when reviewing the PR. I looked at it as a tradeoff between lock contention vs copying the object (connection) and maintaining separate objects in the current implementation. Thought perf tests will help us in picking the right approach. Agree that current approach is becoming confusing and difficult to follow.

Thanks @antoninbas  for proposing the approach that maintains the state in one object and reduces the contention by using the channel. It looks good and definitely has well-defined functionalities for each go routine.
I was wondering if we should do more simpler design first and move to the Antonin's proposed design later in step-by-step basis. Simpler design is where having one object to maintain state (connection store) and have a pointer to the connection in PQ item. Both polling and exporter go routine acquires CS lock to protect the connection store and priority queue objects. I understand there may be lock contention between routines. I feel this way we have more control on different pieces and the performance improvements that we could get. We may need a new perf test. Currently we do not have one that runs both go routines at the same time.