@srikartati 2021-08-16
> Running exporter perf test - BenchmarkExportConntrackConns:
> 
> working branch:
> 
> # conns	iterations	ns/op	B/op	allocs/op
> 20000	13935	1718657	7666	123
> 30000	13421	2497823	8505	144
> 40000	8895	1180317	14332	237
> main branch:
> 
> # conns	iterations	ns/op	B/op	allocs/op
> 20000	143	10320774	1651370	9170
> 30000	76	17695434	1358809	22884
> 40000	35	47754824	9045501	75455
> The results of BenchmarkExportDenyConns are very close to conntrack conn. Can't explain now why ns/op drops from 30000 to 40000.
> 
> Will try to generate a diagram similar to the one above based on @antoninbas 's proposed design. Many thanks to providing the valuable feedback and generating solutions.

@heanlan The improvements with the working branch are remarkable corresponding to all three metrics. Does the working branch mean your current approach? Do we where the improvements are and what were the bottlenecks in the existing code  on the main branch?