@heanlan 2021-08-16
> @heanlan The improvements with the working branch are remarkable corresponding to all three metrics. Does the working branch mean your current approach? Do we where the improvements are and what were the bottlenecks in the existing code on the main branch?

re-run the perf-test with fixed iterations. The results looks much more reasonable.

set-up | # conns | iterations | ns/op | B/op | allocs/op
-- | -- | -- | -- | -- | --
current approach | 20000 | 100 | 3264823 | 421728 | 7679
main branch | 20000 | 100 | 9586789 | 464441 | 8009