@heanlan 2021-08-16
Thanks @srikartati for the comments and generating the simpler design, a diagram - [link](https://drive.google.com/file/d/1JONrhkZPHQM9IzwCuGVBCtlQMiq6NtsR/view?usp=sharing)
I have listed the changes from the former one in red on the side. 

Also made a small change to @antoninbas 's design - by adding `connKey` as a field in `ItemToExpire`, as in go routine B, we need to use pqItem to get the corresponding connection. updated diagram - [link](https://drive.google.com/file/d/1Dga2HRm-GaQB0n1nRIRUgJbZ1Wmm4KYq/view?usp=sharing)